+++ dave/server/debuggers/gdb_/__init__.py
@@ -5,6 +5,7 @@ st.SERVER_TYPE = st.ServerType.GDB
 
 from ...languages import *
 from .commands import GdbCommand, exit_handler, stop_handler
+from .formatters import dave_printer
 
 
 import gdb  # type: ignore
@@ -14,5 +15,6 @@ import os
 GdbCommand()
 gdb.events.exited.connect(exit_handler)
 gdb.events.stop.connect(stop_handler)
+gdb.pretty_printers.append(dave_printer)
 
 gdb.write("INFO : [dave] Successfully loaded\n")