+++ dave/server/debuggers/value.py
@@ -1,5 +1,11 @@
 from __future__ import annotations
 from abc import ABC, abstractmethod
+from enum import Enum
+
+
+class Endianness(Enum):
+    LITTLE = "<"
+    BIG = ">"
 
 
 class AbstractValue(ABC):
@@ -14,10 +20,6 @@ class AbstractValue(ABC):
     @abstractmethod
     def attr(self, name: str) -> AbstractValue:
         pass
-
-    # @abstractmethod
-    # def call_method(self, name: str, *args) -> AbstractValue:
-    #     pass
 
     @abstractmethod
     def address(self) -> int: