+++ dave/server/languages/c_cpp/choc.py
@@ -6,7 +6,6 @@ from typing import Callable, List, Tuple
 
 from dave.common.logger import Logger
 from ...container import SampleType, Container2D, Container1D
-from ...container_factory import ContainerFactory
 from ...debuggers.value import AbstractValue
 
 
@@ -248,9 +247,9 @@ class ChocInterleavedBuffer(Container2D):
         return self.__view.read_from_debugger()
 
 
-ContainerFactory().register(ChocMonoView)
-ContainerFactory().register(ChocMonoBuffer)
-ContainerFactory().register(ChocChannelArrayView)
-ContainerFactory().register(ChocChannelArrayBuffer)
-ContainerFactory().register(ChocInterleavedView)
-ContainerFactory().register(ChocInterleavedBuffer)
+ChocMonoView.register()
+ChocMonoBuffer.register()
+ChocChannelArrayView.register()
+ChocChannelArrayBuffer.register()
+ChocInterleavedView.register()
+ChocInterleavedBuffer.register()